feat: adapt container style
