+++ test/cases/container-style/comp1.html
@@ -11,7 +11,7 @@
   </style>
   <div>count:{{count}}</div>
   <div>
-    <button on:click="count += 100">count += 100</button>
+    <button on:click="haha">count += 100</button>
   </div>
   <div class="main">
     <slot></slot>
@@ -21,6 +21,13 @@
       tag: "comp-one",
       data: {
         count: 0,
+      },
+      proto: {
+        haha() {
+          this.count += 100;
+
+          $.checkMatch();
+        },
       },
     };
   </script>
